Split skippy stage control module
