+++ crates/mesh-llm/src/inference/skippy/stage/mod.rs
@@ -1,255 +1,21 @@
-use std::{
-    collections::HashMap,
-    net::SocketAddr,
-    path::{Path, PathBuf},
-    sync::Arc,
-    time::Duration,
-};
+use std::{collections::HashMap, net::SocketAddr, sync::Arc, time::Duration};
 
 use anyhow::{anyhow, Context, Result};
-use skippy_protocol::{FlashAttentionType, LoadMode, PeerConfig, StageConfig, StageDevice};
+use skippy_protocol::{FlashAttentionType, LoadMode, PeerConfig, StageConfig};
 use skippy_server::{
     binary_transport::{BinaryStageOptions, WireCondition},
     telemetry::TelemetryLevel,
     EmbeddedServerHandle,
 };
-use tokio::sync::{mpsc, oneshot, Mutex};
+use tokio::sync::{mpsc, Mutex};
 
-use super::materialization::{inspect_stage_package, is_layer_package_ref};
+mod inventory;
+#[cfg(test)]
+mod tests;
+mod types;
 
-#[derive(Debug)]
-pub(crate) struct StageControlCommand {
-    pub(crate) request: StageControlRequest,
-    pub(crate) resp: oneshot::Sender<Result<StageControlResponse>>,
-}
-
-#[derive(Clone, Debug)]
-#[allow(clippy::large_enum_variant)]
-pub(crate) enum StageControlRequest {
-    Load(StageLoadRequest),
-    Stop(StageStopRequest),
-    Status(StageStatusFilter),
-    Inventory(StageInventoryRequest),
-    Prepare(StagePrepareRequest),
-    CancelPrepare(StageCancelPrepareRequest),
-    StatusUpdate(StagePreparationStatus),
-}
-
-#[derive(Clone, Debug)]
-#[allow(clippy::large_enum_variant)]
-pub(crate) enum StageControlResponse {
-    Ready(StageReadyResponse),
-    Status(Vec<StageStatusSnapshot>),
-    Inventory(StageLayerInventory),
-    PrepareAccepted(StagePrepareAcceptedResponse),
-    PreparationStatus(StagePreparationStatus),
-    StatusAck(StageStatusAck),
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StageLoadRequest {
-    pub(crate) topology_id: String,
-    pub(crate) run_id: String,
-    pub(crate) model_id: String,
-    pub(crate) backend: String,
-    pub(crate) package_ref: String,
-    pub(crate) manifest_sha256: String,
-    pub(crate) stage_id: String,
-    pub(crate) stage_index: u32,
-    pub(crate) layer_start: u32,
-    pub(crate) layer_end: u32,
-    pub(crate) model_path: Option<String>,
-    pub(crate) projector_path: Option<String>,
-    pub(crate) selected_device: Option<StageDevice>,
-    pub(crate) bind_addr: String,
-    pub(crate) activation_width: i32,
-    pub(crate) wire_dtype: StageWireDType,
-    pub(crate) ctx_size: u32,
-    pub(crate) lane_count: u32,
-    pub(crate) n_batch: Option<u32>,
-    pub(crate) n_ubatch: Option<u32>,
-    pub(crate) n_gpu_layers: i32,
-    pub(crate) cache_type_k: String,
-    pub(crate) cache_type_v: String,
-    pub(crate) flash_attn_type: FlashAttentionType,
-    pub(crate) shutdown_generation: u64,
-    pub(crate) load_mode: LoadMode,
-    pub(crate) upstream: Option<StagePeerDescriptor>,
-    pub(crate) downstream: Option<StagePeerDescriptor>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StageStopRequest {
-    pub(crate) topology_id: String,
-    pub(crate) run_id: String,
-    pub(crate) stage_id: String,
-    pub(crate) shutdown_generation: u64,
-}
-
-#[derive(Clone, Debug, Default)]
-pub(crate) struct StageStatusFilter {
-    pub(crate) topology_id: Option<String>,
-    pub(crate) run_id: Option<String>,
-    pub(crate) stage_id: Option<String>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StageInventoryRequest {
-    pub(crate) model_id: String,
-    pub(crate) package_ref: String,
-    pub(crate) manifest_sha256: String,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StagePrepareRequest {
-    pub(crate) load: StageLoadRequest,
-    pub(crate) coordinator_id: Option<iroh::EndpointId>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StageCancelPrepareRequest {
-    pub(crate) topology_id: String,
-    pub(crate) run_id: String,
-    pub(crate) stage_id: String,
-    pub(crate) shutdown_generation: u64,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct LayerRange {
-    pub(crate) layer_start: u32,
-    pub(crate) layer_end: u32,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StageLayerInventory {
-    pub(crate) model_id: String,
-    pub(crate) package_ref: String,
-    pub(crate) manifest_sha256: String,
-    pub(crate) layer_count: u32,
-    pub(crate) ready_ranges: Vec<LayerRange>,
-    pub(crate) available_ranges: Vec<LayerRange>,
-    pub(crate) missing_ranges: Vec<LayerRange>,
-    pub(crate) preparing_ranges: Vec<StagePreparationStatus>,
-    pub(crate) source_model_path: Option<String>,
-    pub(crate) source_model_bytes: Option<u64>,
-    pub(crate) source_model_kind: SourceModelKind,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum SourceModelKind {
-    Unknown,
-    LayerPackage,
-    PlainGguf,
-    SplitGguf,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StagePeerDescriptor {
-    pub(crate) stage_id: String,
-    pub(crate) stage_index: u32,
-    pub(crate) endpoint: String,
-    pub(crate) node_id: Option<iroh::EndpointId>,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum StageWireDType {
-    F32,
-    F16,
-    Q8,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum StageRuntimeState {
-    Starting,
-    Ready,
-    Stopping,
-    Stopped,
-    Failed,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum StagePreparationState {
-    Assigned,
-    Downloading,
-    Available,
-    Resolving,
-    Loading,
-    Ready,
-    Failed,
-    Cancelled,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StageReadyResponse {
-    pub(crate) accepted: bool,
-    pub(crate) status: StageStatusSnapshot,
-    pub(crate) error: Option<String>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StageStatusSnapshot {
-    pub(crate) topology_id: String,
-    pub(crate) run_id: String,
-    pub(crate) model_id: String,
-    pub(crate) backend: String,
-    pub(crate) package_ref: Option<String>,
-    pub(crate) manifest_sha256: Option<String>,
-    pub(crate) source_model_path: Option<String>,
-    pub(crate) source_model_sha256: Option<String>,
-    pub(crate) source_model_bytes: Option<u64>,
-    pub(crate) materialized_path: Option<String>,
-    pub(crate) materialized_pinned: bool,
-    pub(crate) projector_path: Option<String>,
-    pub(crate) stage_id: String,
-    pub(crate) stage_index: u32,
-    pub(crate) layer_start: u32,
-    pub(crate) layer_end: u32,
-    pub(crate) state: StageRuntimeState,
-    pub(crate) bind_addr: String,
-    pub(crate) activation_width: u32,
-    pub(crate) wire_dtype: StageWireDType,
-    pub(crate) selected_device: Option<StageDevice>,
-    pub(crate) ctx_size: u32,
-    pub(crate) lane_count: u32,
-    pub(crate) n_batch: Option<u32>,
-    pub(crate) n_ubatch: Option<u32>,
-    pub(crate) flash_attn_type: FlashAttentionType,
-    pub(crate) error: Option<String>,
-    pub(crate) shutdown_generation: u64,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StagePreparationStatus {
-    pub(crate) topology_id: String,
-    pub(crate) run_id: String,
-    pub(crate) model_id: String,
-    pub(crate) backend: String,
-    pub(crate) package_ref: String,
-    pub(crate) manifest_sha256: String,
-    pub(crate) stage_id: String,
-    pub(crate) stage_index: u32,
-    pub(crate) layer_start: u32,
-    pub(crate) layer_end: u32,
-    pub(crate) state: StagePreparationState,
-    pub(crate) bytes_done: Option<u64>,
-    pub(crate) bytes_total: Option<u64>,
-    pub(crate) bind_addr: Option<String>,
-    pub(crate) error: Option<String>,
-    pub(crate) shutdown_generation: u64,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StagePrepareAcceptedResponse {
-    pub(crate) accepted: bool,
-    pub(crate) status: StagePreparationStatus,
-    pub(crate) error: Option<String>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) struct StageStatusAck {
-    pub(crate) accepted: bool,
-    pub(crate) error: Option<String>,
-}
+use inventory::{resolve_inventory_source, run_stage_prepare_task};
+pub(crate) use types::*;
 
 struct RunningStage {
     load: StageLoadRequest,
@@ -502,151 +268,6 @@ impl StageControlState {
             .map(status_from_running)
             .collect()
     }
-}
-
-#[derive(Clone, Debug)]
-struct InventorySource {
-    path: PathBuf,
-    bytes: Option<u64>,
-    layer_count: u32,
-    kind: SourceModelKind,
-}
-
-fn resolve_inventory_source(request: &StageInventoryRequest) -> Option<InventorySource> {
-    if is_layer_package_ref(&request.package_ref) {
-        let info = inspect_stage_package(&request.package_ref).ok()?;
-        return Some(InventorySource {
-            path: info.package_dir,
-            bytes: info.source_model_bytes,
-            layer_count: info.layer_count,
-            kind: SourceModelKind::LayerPackage,
-        });
-    }
-
-    for candidate in inventory_source_candidates(request) {
-        if !candidate.exists() {
-            continue;
-        }
-        let layer_count = super::infer_layer_count(&candidate).ok()?;
-        let kind = if is_split_gguf_path(&candidate) {
-            SourceModelKind::SplitGguf
-        } else {
-            SourceModelKind::PlainGguf
-        };
-        let bytes = crate::inference::election::total_model_bytes(&candidate);
-        return Some(InventorySource {
-            path: candidate,
-            bytes: Some(bytes),
-            layer_count,
-            kind,
-        });
-    }
-    None
-}
-
-fn inventory_source_candidates(request: &StageInventoryRequest) -> Vec<PathBuf> {
-    let mut candidates = Vec::new();
-    if let Some(path) = request.package_ref.strip_prefix("gguf://") {
-        if !path.is_empty() {
-            candidates.push(PathBuf::from(path));
-        }
-    }
-    if !request.model_id.is_empty() {
-        candidates.push(crate::models::find_model_path(&request.model_id));
-    }
-    candidates
-}
-
-fn is_split_gguf_path(path: &Path) -> bool {
-    path.file_name()
-        .and_then(|name| name.to_str())
-        .and_then(model_ref::split_gguf_shard_info)
-        .is_some()
-}
-
-async fn run_stage_prepare_task(
-    preparations: Arc<Mutex<HashMap<String, StagePreparationStatus>>>,
-    key: String,
-    load: StageLoadRequest,
-) {
-    update_preparation(
-        &preparations,
-        &key,
-        preparation_status_from_load(&load, StagePreparationState::Resolving),
-    )
-    .await;
-    if load.load_mode != LoadMode::LayerPackage && !is_layer_package_ref(&load.package_ref) {
-        update_preparation(
-            &preparations,
-            &key,
-            preparation_status_from_load(&load, StagePreparationState::Downloading),
-        )
-        .await;
-    }
-    let result = prepare_stage_source(&load).await;
-    let state = match result {
-        Ok(PrepareSourceResult { bytes_total }) => {
-            let mut status = preparation_status_from_load(&load, StagePreparationState::Available);
-            status.bytes_done = bytes_total;
-            status.bytes_total = bytes_total;
-            status
-        }
-        Err(error) => {
-            let mut status = preparation_status_from_load(&load, StagePreparationState::Failed);
-            status.error = Some(error.to_string());
-            status
-        }
-    };
-    update_preparation(&preparations, &key, state).await;
-}
-
-struct PrepareSourceResult {
-    bytes_total: Option<u64>,
-}
-
-async fn prepare_stage_source(load: &StageLoadRequest) -> Result<PrepareSourceResult> {
-    if load.load_mode == LoadMode::LayerPackage || is_layer_package_ref(&load.package_ref) {
-        let info = inspect_stage_package(&load.package_ref)?;
-        return Ok(PrepareSourceResult {
-            bytes_total: info.source_model_bytes,
-        });
-    }
-
-    for candidate in [
-        load.model_path.as_deref(),
-        Some(load.model_id.as_str()),
-        load.package_ref.strip_prefix("gguf://"),
-    ]
-    .into_iter()
-    .flatten()
-    .filter(|candidate| !candidate.is_empty())
-    {
-        match crate::models::resolve_model_spec_with_progress(Path::new(candidate), false).await {
-            Ok(path) => {
-                let bytes_total = crate::inference::election::total_model_bytes(&path);
-                return Ok(PrepareSourceResult {
-                    bytes_total: Some(bytes_total),
-                });
-            }
-            Err(last_error) => {
-                tracing::debug!(
-                    stage_id = %load.stage_id,
-                    candidate,
-                    error = %last_error,
-                    "stage source prepare candidate failed"
-                );
-            }
-        }
-    }
-    anyhow::bail!("stage source model is not available")
-}
-
-async fn update_preparation(
-    preparations: &Arc<Mutex<HashMap<String, StagePreparationStatus>>>,
-    key: &str,
-    status: StagePreparationStatus,
-) {
-    preparations.lock().await.insert(key.to_string(), status);
 }
 
 impl StageStatusFilter {
@@ -937,202 +558,5 @@ impl From<StageWireDType> for skippy_protocol::binary::WireActivationDType {
             StageWireDType::F16 => Self::F16,
             StageWireDType::Q8 => Self::Q8,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::time::Instant;
-
-    fn load_request() -> StageLoadRequest {
-        StageLoadRequest {
-            topology_id: "topology-a".to_string(),
-            run_id: "run-a".to_string(),
-            model_id: "model-a".to_string(),
-            backend: "skippy".to_string(),
-            package_ref: "pkg-a".to_string(),
-            manifest_sha256: "sha256".to_string(),
-            stage_id: "stage-0".to_string(),
-            stage_index: 0,
-            layer_start: 0,
-            layer_end: 12,
-            model_path: Some("/models/model.gguf".to_string()),
-            projector_path: Some("/models/mmproj.gguf".to_string()),
-            selected_device: Some(StageDevice {
-                backend_device: "CUDA0".to_string(),
-                stable_id: Some("GPU-123".to_string()),
-                index: Some(0),
-                vram_bytes: Some(24_000_000_000),
-            }),
-            bind_addr: "127.0.0.1:0".to_string(),
-            activation_width: 4096,
-            wire_dtype: StageWireDType::F16,
-            ctx_size: 8192,
-            lane_count: 3,
-            n_batch: Some(2048),
-            n_ubatch: Some(512),
-            n_gpu_layers: -1,
-            cache_type_k: "f16".to_string(),
-            cache_type_v: "q8_0".to_string(),
-            flash_attn_type: FlashAttentionType::Enabled,
-            shutdown_generation: 7,
-            load_mode: LoadMode::RuntimeSlice,
-            upstream: None,
-            downstream: Some(StagePeerDescriptor {
-                stage_id: "stage-1".to_string(),
-                stage_index: 1,
-                endpoint: "127.0.0.1:9001".to_string(),
-                node_id: None,
-            }),
-        }
-    }
-
-    #[test]
-    fn stage_config_preserves_backend_neutral_load_fields() {
-        let request = load_request();
-        let config = stage_config(&request, None).unwrap();
-
-        assert_eq!(config.topology_id, "topology-a");
-        assert_eq!(config.run_id, "run-a");
-        assert_eq!(config.model_id, "model-a");
-        assert_eq!(config.package_ref.as_deref(), Some("pkg-a"));
-        assert_eq!(config.manifest_sha256.as_deref(), Some("sha256"));
-        assert_eq!(
-            config.source_model_path.as_deref(),
-            Some("/models/model.gguf")
-        );
-        assert!(config.materialized_path.is_none());
-        assert!(!config.materialized_pinned);
-        assert_eq!(config.stage_id, "stage-0");
-        assert_eq!(config.stage_index, 0);
-        assert_eq!(config.layer_start, 0);
-        assert_eq!(config.layer_end, 12);
-        assert_eq!(config.lane_count, 3);
-        assert_eq!(config.n_batch, Some(2048));
-        assert_eq!(config.n_ubatch, Some(512));
-        assert_eq!(config.model_path.as_deref(), Some("/models/model.gguf"));
-        assert_eq!(
-            config.projector_path.as_deref(),
-            Some("/models/mmproj.gguf")
-        );
-        assert_eq!(config.flash_attn_type, FlashAttentionType::Enabled);
-        assert_eq!(
-            config
-                .selected_device
-                .as_ref()
-                .map(|d| d.backend_device.as_str()),
-            Some("CUDA0")
-        );
-        assert_eq!(
-            config.downstream.as_ref().map(|d| d.stage_id.as_str()),
-            Some("stage-1")
-        );
-        assert!(config.filter_tensors_on_load);
-    }
-
-    #[test]
-    fn stage_config_rejects_empty_selected_backend_device() {
-        let mut request = load_request();
-        request.selected_device = Some(StageDevice {
-            backend_device: String::new(),
-            stable_id: Some("uuid:GPU-123".into()),
-            index: Some(0),
-            vram_bytes: Some(24_000_000_000),
-        });
-
-        let err = stage_config(&request, None).unwrap_err().to_string();
-
-        assert!(err.contains("selected backend device"));
-    }
-
-    #[test]
-    fn stage_status_filter_matches_optional_identity_fields() {
-        let load = load_request();
-        assert!(StageStatusFilter {
-            topology_id: Some("topology-a".to_string()),
-            run_id: None,
-            stage_id: Some("stage-0".to_string()),
-        }
-        .matches(&load));
-        assert!(!StageStatusFilter {
-            topology_id: Some("other".to_string()),
-            run_id: None,
-            stage_id: None,
-        }
-        .matches(&load));
-    }
-
-    #[test]
-    fn materialize_stage_bind_addr_replaces_ephemeral_port() {
-        let bind_addr = materialize_stage_bind_addr("127.0.0.1:0".parse().unwrap()).unwrap();
-        assert_eq!(bind_addr.ip().to_string(), "127.0.0.1");
-        assert_ne!(bind_addr.port(), 0);
-    }
-
-    #[tokio::test]
-    async fn binary_stage_ready_probe_waits_for_wire_handshake() {
-        let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
-        let bind_addr = listener.local_addr().unwrap();
-        let server = std::thread::spawn(move || {
-            std::thread::sleep(Duration::from_millis(75));
-            let (mut stream, _) = listener.accept().unwrap();
-            skippy_protocol::binary::send_ready(&mut stream).unwrap();
-        });
-
-        let started = Instant::now();
-        wait_for_binary_stage_ready(bind_addr, Duration::from_secs(2))
-            .await
-            .unwrap();
-        assert!(started.elapsed() >= Duration::from_millis(50));
-        server.join().unwrap();
-    }
-
-    #[tokio::test]
-    async fn prepare_stage_records_background_source_availability() {
-        let file = tempfile::NamedTempFile::new().unwrap();
-        let path = file.path().to_string_lossy().to_string();
-        let mut load = load_request();
-        load.model_path = Some(path.clone());
-        load.package_ref = "gguf:///definitely/missing/model.gguf".to_string();
-        load.downstream = None;
-        let mut state = StageControlState::default();
-
-        let accepted = state
-            .prepare(StagePrepareRequest {
-                load: load.clone(),
-                coordinator_id: None,
-            })
-            .await
-            .unwrap();
-
-        assert!(accepted.accepted);
-        assert_eq!(accepted.status.state, StagePreparationState::Assigned);
-
-        let mut last_state = StagePreparationState::Assigned;
-        for _ in 0..20 {
-            let inventory = state
-                .inventory(StageInventoryRequest {
-                    model_id: load.model_id.clone(),
-                    package_ref: load.package_ref.clone(),
-                    manifest_sha256: load.manifest_sha256.clone(),
-                })
-                .await;
-            if let Some(status) = inventory
-                .preparing_ranges
-                .iter()
-                .find(|status| status.stage_id == load.stage_id)
-            {
-                last_state = status.state;
-                if status.state == StagePreparationState::Available {
-                    assert_eq!(status.bytes_done, Some(0));
-                    assert_eq!(status.bytes_total, Some(0));
-                    return;
-                }
-            }
-            tokio::time::sleep(Duration::from_millis(25)).await;
-        }
-
-        panic!("prepare did not become available, last state: {last_state:?}");
     }
 }
